began implementing fixtures
implemented pool creation, decay, and stake unit tests

 implemented scenarios test in ts

minor format changes

implemented cancelPendingWithdraw function

abi update

renamed tests by fixture order

added new test files

installed solcover and modified fixtures to use typechain for solcover compatibility

added additional stake and initial conditions tests

installed solhint and prettier

added config files for slither and eslint and updated package.json

refactored decay calculation logic in tests

generated interfaces

implemented unstake and withdrawstake tests

implemented failedTest fixture

added AntePoolTest from justin's branch

replaced all instances of supporter with staker

added unit tests for all launch mainnet ante tests

refactored and added new fixture for claim testing

implemented checkTest tests

implemented claim unit tests

implemented antepoolfactory unit tests

+++ contracts/examples/AnteAVLDropTest.sol
@@ -20,27 +20,27 @@ contract AnteAVLDropTest is AnteTest("Ante Doesnt Lose 99% of its AVL") {
 
     uint256 public totalAVL;
     uint256 public avlThreshold;
-    
+
     // Array of contract addresses to test should be passed in when creating
-    constructor (address[] memory _testedContracts) {
+    constructor(address[] memory _testedContracts) {
         protocolName = "Ante Finance";
         testedContracts = _testedContracts;
 
         // Loop over address array and sum up total AVL
-        for (uint i = 0; i < testedContracts.length; i++) {
+        for (uint256 i = 0; i < testedContracts.length; i++) {
             totalAVL = totalAVL.add(testedContracts[i].balance);
         }
-        
+
         // Calculate test failure threshold using 99% drop in total AVL
         avlThreshold = totalAVL.div(100);
     }
-    
+
     function checkTestPasses() public view override returns (bool) {
         uint256 currentAVL;
 
         // Sum up current AVL across tested contracts
         // TODO: to exclude failed tests or no?
-        for (uint i = 0; i < testedContracts.length; i++) {
+        for (uint256 i = 0; i < testedContracts.length; i++) {
             currentAVL = currentAVL.add(testedContracts[i].balance);
         }
